completing the authentication and permission part, will improve how it is done in the future. Now only the owner of the post can make changes to that perticular post like delete, update. And the owner of the post is decided to be the one who is logged in at the time of creating the post

+++ CRUD_api/posts.py
@@ -26,7 +26,7 @@ class Posts:
             self.db.rollback()
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail = "Posts could not be found")
 
-    def create_post(self,post : Post):
+    def create_post(self,post):
         post_dict = post.dict()
 
         if not users.login.user_logged_in():
@@ -37,9 +37,16 @@ class Posts:
                         (post_dict['title'],post_dict['content'],post_dict['publish']))  # %s to prevent SQL injection attacks
             
             data = self.cursor.fetchall()[0]
+            
 
-            self.cursor.execute("UPDATE users SET posts = array_append(posts, %s) WHERE email_id = %s",
+
+            self.cursor.execute("UPDATE users SET posts = array_append(posts, %s) WHERE email_id = %s RETURNING *",
                                  (data['id'],users.login.user_logged_in()))
+            
+            user_data = self.cursor.fetchall()[0]
+            
+            self.cursor.execute("UPDATE posts SET posted_by = %s WHERE id = %s", (user_data['user_id'], data['id']))
+
             
             self.db.commit()
             return {"Post created":post_dict}
@@ -64,17 +71,49 @@ class Posts:
             raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail="You need to be logged in")
         
         try:
+            self.cursor.execute('SELECT posted_by FROM posts WHERE id = %s', (id,))
+            data = self.cursor.fetchall()
+            if data:
+               data = data[0]
+            else:
+                raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+            
+            self.cursor.execute('SELECT email_id FROM users WHERE user_id = %s', (data['posted_by'],))
+            user_data  = self.cursor.fetchall()
+
+            if user_data:
+               user_data = user_data[0]
+            else:
+                self.db.rollback()
+                raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail='You can only delete your own posts')
+            
+            self.db.commit()
+            if user_data['email_id'] != users.login.user_logged_in():
+                self.db.rollback()
+                raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail='You can only delete your own posts')
+            
             self.cursor.execute("DELETE FROM posts WHERE id = %s;",(id,))
             self.db.commit()
             return {"message":"post deleted"}
         except Exception as error:
             self.db.rollback()
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail= error)
-
-    def update_posts(self,id : int, post : Post):
+            raise error
+        
+    def update_posts(self,id : int, post):
         if not users.login.user_logged_in():
             raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail="You need to be logged in")
+
         try:
+            
+            self.cursor.execute('SELECT posted_by FROM posts WHERE id = %s', (id,))
+            data = self.cursor.fetchall()[0]
+            self.cursor.execute('SELECT email_id FROM users WHERE user_id = %s', (data['posted_by'],))
+            user_data  = self.cursor.fetchall()[0]
+            self.db.commit()
+            if user_data['email_id'] != users.login.user_logged_in():
+                self.db.rollback()
+                raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail='You can only update your own posts')
+            
             post = post.dict()
             self.cursor.execute("UPDATE posts SET title=%s, content = %s,publish = %s WHERE id = %s RETURNING *;",
                             (post['title'], post['content'],post['publish'],id))
@@ -84,4 +123,4 @@ class Posts:
         except Exception as error:
             print(error)
             self.db.rollback()
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Post could not be found")
+            raise error
